api/requirements.txt
google-auth==2.25.2
google-auth-oauthlib==1.2.0
requests==2.31.0
